Bai cannot be deleted when TransferInfos still depend on it.

+++ src/main/java/org/bait/service/BaiServiceImpl.java
@@ -3,11 +3,16 @@ package org.bait.service;
 import org.bait.db.BaiRepository;
 import org.bait.db.model.BaiDbImpl;
 import org.bait.model.Bai;
+import org.bait.model.TransferInfo;
 import org.bait.service.api.BaiService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Required;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Collection;
+import java.util.LinkedList;
+import java.util.List;
 
 @Service
 public class BaiServiceImpl implements BaiService {
@@ -40,7 +45,16 @@ public class BaiServiceImpl implements BaiService {
 
     @Override
     @Transactional
-    public void deleteBankAccountInformation(final String baiId) {
+    public void deleteBankAccountInformation(final String baiId) throws PreconditionFailedException {
+        BaiDbImpl bai = baiRepository.findOne(baiId);
+        Collection<TransferInfo> transferInfoForBai = bai.getTransferInfo();
+        if (transferInfoForBai.size() >= 1) {
+            List<String> transferIds = new LinkedList<>();
+            for (TransferInfo transferInfo : transferInfoForBai) {
+                transferIds.add(transferInfo.getTransferId());
+            }
+            throw new PreconditionFailedException(String.format("Bai cannot be deleted. Transfer object '%s' depend upon it. Delete them first in order to delete Bai.", transferIds));
+        }
         baiRepository.deleteByBaiId(baiId);
     }
